[!!!][TASK] Add EventLog support
This change add support for EventLog, based on Neos EventLog so currently
this package is dependent on TYPO3.Neos. Maybe in the future we can have a
decoupled package for the EventLog. There's some hack to persist Event in
the current order, because we use sub command in the CLI controller.

This change is marked has breaking because the signature of the
Importer::skipNodeProcessing changed and you need to update your Importer
to follow the change in ImporterInterface. It's a best practice to not
implement directly the ImporterInterface, extend the abstract Importer.