!!! FEATURE: Make settings structure more extensible
This change moves the "parts" of the import settings into a sub-array
called "parts". Previously, they were just a direct child of the respective
preset array.

This structure change allows for defining further options per preset
in the future. For example, we might introduce tasks which can be
configured to be executed before or after an import, and thus belong
to the preset and not to an individual part.

If you are using the settings structure which was valid before this change,
you'll get a meaningful error and explanation when running the import
command.