FEATURE: Prevent multiple import runs based on reference
You may optionally specify an external import identifier when running
the import:batch command which will be stored as meta data with the
import run. This identifier is used for checking if an import of the
given data set (in general) has been executed earlier.

The external import identifier therefore does globally what the external
record identifier does on a per record basis.

If an external import identifier was specified and an import using that
identifier has been executed earlier, the import command will stop with
a corresponding message. You can force running such an import by
specifying the new --force flag.