TASK: Add option to enable/disable creating event log entries for records
Currently, if the Neos Event Log is active, besides general information about
the import run or the preset/part currently processed, the importer writes
three records to the event log for each item imported. If you have a lot of
records and/or the import is running frequently, this will make the eventlog
table growing quickly.

In case you have sufficient logging without the event log, you might want to
disable logging on a per-record basis. This change adds a new setting without
changing the default functionality. Refer to the README for more information.